Reuse code UploadButton
